requirements.txt
datasets
transformers
torchtune
torchao
matplotlib
optuna
# lm-eval

# Single T4 GPU training
# tensorstore
# zarr

# Note: Some experiments have additional requirements
# See individual experiment directories for requirements.txt